+++ src/GraphQl/Serializer/SerializerContextBuilder.php
@@ -16,6 +16,7 @@ namespace ApiPlatform\Core\GraphQl\Serializer;
 use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
 use ApiPlatform\Core\Metadata\Resource\ResourceMetadata;
 use GraphQL\Type\Definition\ResolveInfo;
+use Symfony\Component\Serializer\NameConverter\AdvancedNameConverterInterface;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 
 /**
@@ -45,10 +46,6 @@ final class SerializerContextBuilder implements SerializerContextBuilderInterfac
             'graphql_operation_name' => $operationName,
         ];
 
-        if ($normalization) {
-            $context['attributes'] = $this->fieldsToAttributes($resourceMetadata, $resolverContext);
-        }
-
         if (isset($resolverContext['fields'])) {
             $context['no_resolver_data'] = true;
         }
@@ -61,25 +58,29 @@ final class SerializerContextBuilder implements SerializerContextBuilderInterfac
             $context = array_merge($resourceMetadata->getGraphqlAttribute($operationName, $key, [], true), $context);
         }
 
+        if ($normalization) {
+            $context['attributes'] = $this->fieldsToAttributes($resourceClass, $resourceMetadata, $resolverContext, $context);
+        }
+
         return $context;
     }
 
     /**
      * Retrieves fields, recursively replaces the "_id" key (the raw id) by "id" (the name of the property expected by the Serializer) and flattens edge and node structures (pagination).
      */
-    private function fieldsToAttributes(?ResourceMetadata $resourceMetadata, array $context): array
+    private function fieldsToAttributes(?string $resourceClass, ?ResourceMetadata $resourceMetadata, array $resolverContext, array $context): array
     {
-        if (isset($context['fields'])) {
-            $fields = $context['fields'];
+        if (isset($resolverContext['fields'])) {
+            $fields = $resolverContext['fields'];
         } else {
             /** @var ResolveInfo $info */
-            $info = $context['info'];
+            $info = $resolverContext['info'];
             $fields = $info->getFieldSelection(PHP_INT_MAX);
         }
 
-        $attributes = $this->replaceIdKeys($fields['edges']['node'] ?? $fields);
+        $attributes = $this->replaceIdKeys($fields['edges']['node'] ?? $fields, $resourceClass, $context);
 
-        if ($context['is_mutation'] || $context['is_subscription']) {
+        if ($resolverContext['is_mutation'] || $resolverContext['is_subscription']) {
             if (!$resourceMetadata) {
                 throw new \LogicException('ResourceMetadata should always exist for a mutation or a subscription.');
             }
@@ -92,7 +93,7 @@ final class SerializerContextBuilder implements SerializerContextBuilderInterfac
         return $attributes;
     }
 
-    private function replaceIdKeys(array $fields): array
+    private function replaceIdKeys(array $fields, ?string $resourceClass, array $context): array
     {
         $denormalizedFields = [];
 
@@ -103,14 +104,21 @@ final class SerializerContextBuilder implements SerializerContextBuilderInterfac
                 continue;
             }
 
-            $denormalizedFields[$this->denormalizePropertyName((string) $key)] = \is_array($fields[$key]) ? $this->replaceIdKeys($fields[$key]) : $value;
+            $denormalizedFields[$this->denormalizePropertyName((string) $key, $resourceClass, $context)] = \is_array($fields[$key]) ? $this->replaceIdKeys($fields[$key], $resourceClass, $context) : $value;
         }
 
         return $denormalizedFields;
     }
 
-    private function denormalizePropertyName(string $property): string
+    private function denormalizePropertyName(string $property, ?string $resourceClass, array $context): string
     {
-        return null !== $this->nameConverter ? $this->nameConverter->denormalize($property) : $property;
+        if (null === $this->nameConverter) {
+            return $property;
+        }
+        if ($this->nameConverter instanceof AdvancedNameConverterInterface) {
+            return $this->nameConverter->denormalize($property, $resourceClass, null, $context);
+        }
+
+        return $this->nameConverter->denormalize($property);
     }
 }